Example instance directory
This is an example of what is/was in our instance directory which have
now been moved out to our vucoeus.instance repo. Hopefully this will
help someone who tries to build and run the project locally.

The contents of instance.example/PROD-example come from our PROD
configuration. The only change is that I have remvoed some unused files
and have changed all credentials to placeholder entries. All of the
placeholders start with '<<' and end with '>>'.